handle bad credentials and move logic to auth folder
