+++ src/app/(auth)/login/page.tsx
@@ -1,21 +1,5 @@
-import { AuthService } from "@/auth/auth-service";
+import { LoginPage as LoginForm } from "@/auth/components/login-page";
 
 export default function LoginPage() {
-  return (
-    <form
-      action={async (formData) => {
-        "use server";
-
-        const authService = new AuthService();
-        const username = formData.get("username");
-        const password = formData.get("password");
-
-        await authService.login(username as string, password as string);
-      }}
-    >
-      <input type="text" name="username" id="username" />
-      <input type="password" name="password" id="password" />
-      <button type="submit">Login</button>
-    </form>
-  );
+  return <LoginForm />;
 }
